refactor: move ActivityGateway to Activity package
