+++ src/main/java/com/alkemy/ong/domain/activity/ActivityDefaultGateway.java
@@ -2,7 +2,6 @@ package com.alkemy.ong.domain.activity;
 
 import com.alkemy.ong.data.entity.ActivityEntity;
 import com.alkemy.ong.data.repository.ActivityRepository;
-import com.alkemy.ong.domain.gateway.ActivityGateway;
 import com.alkemy.ong.domain.model.Activity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
