+++ src/main/java/com/alkemy/ong/domain/activity/ActivityService.java
@@ -1,6 +1,5 @@
 package com.alkemy.ong.domain.activity;
 
-import com.alkemy.ong.domain.gateway.ActivityGateway;
 import com.alkemy.ong.domain.model.Activity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
